backend/todo_d/requirements/development.txt
beautifulsoup4==4.5.1
Django==1.10.4
django-appconf==1.0.2
django-compressor==2.1
django-debug-toolbar==1.6
django-environ==0.4.1
django-filter==1.0.1
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.5.3
drf-nested-routers==0.11.1
lxml==3.7.1
Markdown==2.6.7
mongoengine==0.11.0
mysqlclient==1.3.9
pymongo==3.4.0
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.2
